pgsql: Move detector to pgsql/detector module
